hide vacant tags
This patch prevents dwm from drawing tags with no clients (i.e. vacant) on the bar.

It also makes sure that pressing a tag on the bar behaves accordingly by not reserving reactive regions on the bar for vacant tags.

It also stops drawing empty rectangles on the bar for non-vacant tags as there is no need anymore to distinguish vacant tags and it offers a more visible contrast than if there were filled/empty rectangles.